code/requirements.txt
fire==0.4.0
json5==0.9.6
opencv-python==4.5.3.56
pandas==1.1.5
sklearn==0.0
torchnet==0.0.4
tqdm==4.62.3
visdom==0.1.8.9
json
si_hun
